+++ application/ui/src/features/dataset/gallery/toolbar/media-filtering/date-filter/date-filter.component.tsx
@@ -6,9 +6,9 @@ import { getLocalTimeZone, parseAbsoluteToLocal, type DateValue } from '@interna
 import dayjs from 'dayjs';
 import { useDatasetFiltersSearchParams } from 'hooks/use-dataset-filters-search-params.hook';
 
-import { FilterPopoverButton } from '../../../../../components/filter-popover-button/filter-popover-button.component';
-import { formatFilterDate } from '../../../../../shared/date-utils';
-import { isNonEmptyArray } from '../../../../../shared/util';
+import { FilterPopoverButton } from '../../../../../../components/filter-popover-button/filter-popover-button.component';
+import { formatFilterDate } from '../../../../../../shared/date-utils';
+import { isNonEmptyArray } from '../../../../../../shared/util';
 
 const MIN_DATE = parseAbsoluteToLocal(dayjs('2020-01-30').startOf('d').toISOString());
 const MAX_DATE = parseAbsoluteToLocal(dayjs('9999-11-30').endOf('d').toISOString());
