+++ application/ui/src/features/dataset/gallery/toolbar/media-filtering/filter-by-status/filter-by-status.component.tsx
@@ -4,7 +4,7 @@
 import { Item, Picker } from '@geti-ui/ui';
 import { useDatasetFiltersSearchParams } from 'hooks/use-dataset-filters-search-params.hook';
 
-import { FilterByStatusKey } from '../../../../../constants/shared-types';
+import { FilterByStatusKey } from '../../../../../../constants/shared-types';
 
 const FILTER_BY_STATUS_OPTIONS: { name: string; key: FilterByStatusKey }[] = [
     { name: 'All media', key: 'all' },
