+++ application/ui/src/features/dataset/gallery/toolbar/media-filtering/filter-by-subset/filter-by-subset.component.tsx
@@ -4,10 +4,10 @@
 import { useDatasetFiltersSearchParams } from 'hooks/use-dataset-filters-search-params.hook';
 import { isEmpty } from 'lodash-es';
 
-import { FilterPopoverButton } from '../../../../../components/filter-popover-button/filter-popover-button.component';
-import { MultiSelectList } from '../../../../../components/multi-select-list/multi-select-list.component';
-import { DatasetSubset } from '../../../../../constants/shared-types';
-import { pluralize } from '../../../../../shared/util';
+import { FilterPopoverButton } from '../../../../../../components/filter-popover-button/filter-popover-button.component';
+import { MultiSelectList } from '../../../../../../components/multi-select-list/multi-select-list.component';
+import { DatasetSubset } from '../../../../../../constants/shared-types';
+import { pluralize } from '../../../../../../shared/util';
 
 const SUBSET_OPTIONS: { id: DatasetSubset; name: string }[] = [
     { id: 'training', name: 'Training subset' },
