+++ application/ui/src/features/dataset/gallery/toolbar/media-filtering/filter-chips/filter-chips.component.tsx
