+++ application/ui/src/features/dataset/gallery/toolbar/media-filtering/filter-chips/filter-chips.module.scss
