+++ application/ui/src/features/dataset/gallery/toolbar/media-filtering/media-filter-labels/media-filter-labels.component.tsx
@@ -5,9 +5,9 @@ import { useDatasetFiltersSearchParams } from 'hooks/use-dataset-filters-search-
 import { useProjectLabels } from 'hooks/use-project-labels.hook';
 import { isEmpty } from 'lodash-es';
 
-import { FilterPopoverButton } from '../../../../../components/filter-popover-button/filter-popover-button.component';
-import { MultiSelectList } from '../../../../../components/multi-select-list/multi-select-list.component';
-import { pluralize } from '../../../../../shared/util';
+import { FilterPopoverButton } from '../../../../../../components/filter-popover-button/filter-popover-button.component';
+import { MultiSelectList } from '../../../../../../components/multi-select-list/multi-select-list.component';
+import { pluralize } from '../../../../../../shared/util';
 
 export const MediaFilterLabels = () => {
     const labels = useProjectLabels();
